+++ src/Sentry/Laravel/Tracing/Middleware.php
@@ -4,7 +4,6 @@ namespace Sentry\Laravel\Tracing;
 
 use Closure;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 use Illuminate\Routing\Route;
 use Sentry\Laravel\Integration;
 use Sentry\SentrySdk;
@@ -12,6 +11,7 @@ use Sentry\State\HubInterface;
 use Sentry\Tracing\Span;
 use Sentry\Tracing\SpanContext;
 use Sentry\Tracing\TransactionContext;
+use Symfony\Component\HttpFoundation\Response;
 
 class Middleware
 {
@@ -56,8 +56,8 @@ class Middleware
     /**
      * Handle the application termination.
      *
-     * @param \Illuminate\Http\Request  $request
-     * @param \Illuminate\Http\Response $response
+     * @param \Illuminate\Http\Request                   $request
+     * @param \Symfony\Component\HttpFoundation\Response $response
      *
      * @return void
      */
@@ -208,7 +208,7 @@ class Middleware
 
     private function hydrateResponseData(Response $response): void
     {
-        $this->transaction->setHttpStatus($response->status());
+        $this->transaction->setHttpStatus($response->getStatusCode());
     }
 
     private function updateTransactionNameIfDefault(?string $name): void